src/detector/requirements.txt
auto_mix_prep==0.2.0
pyzmq==22.0.3
